RakuAST: handle 'has method' better in packages
Turn the use of 'method foo' in a package type that cannot support
methods, into a worry rather than an execution error during parsing.

Add a class X::Useless::Declaration to handle this type of worry.

Note that the worry is issued on the resolver at the moment, because
when issued on the method object , it will not be seen for some reason.

This sadly means there is no position in the code yet to be shown.
But since this is a worry, we can live with that for now.