RakuAST: remove known attributes from seen attributes
The idea being that we will be able to provide a proper sorry for
those attributes that are referenced in a class, but which do not
exist in that class (which could possibly have consumed one or more
roles, so we need to check *after* composing).

Sadly haven't found a way to actually check for remaining attributes
seen yet.