Fix X::Hash::Store::OddNumber message with undefined last element
First of all, the last element can be an undefined value and we better
report it for ease of debugging. The current fix is incomplete with
regard to `Mu` which is no less valid as an element than any other type
object. But for now let's pretend that it's the least probable value to
appear in this role.

Second, if there are few elements found in initializer and no last one
to be reported, then instead of `$msg` the method was returning an empty
list which is a normal outcome of the last postfix `if`.

Also improve the message itself by reporting that the last element is a
type object when it is one.