Don't set callers $/ with Grammar.parse in 6.e
The fact that Grammar.parse/subparse/parsefile all set $/ in the
caller's scope, is a race-condition waiting to happen.

See discussion started by DarthGandalf at
  https://irclogs.raku.org/raku-beginner/2023-11-27.html#09:09

The problem is really that blocks do *NOT* have their own lexical
$/, wherease subroutines/methods do.  But 6.c/d Grammar.parse are
expected to set the caller's lexical $/.  Which if the code is inside
a block, will set the first lexical $/ found in the call stack.

So in:
    sub foo($input) {
        $input.lines.race.map: { G.parse($line) }
    }
$/ will be set in the scope of "sub foo", and *NOT* in the scope of
{ G.parse($line) }.  Which is normally not much of an issue, unless
the block happens to be part of a "hyper" or a "race" chain.

With this commit $/ will simply *NOT* be set by Grammar.parse/subparse/
parsefile anymore.  This feels a lot cleaner than having to explain
when (implicitely) using $/ is a good thing or not.

However this does *NOT* fix the issue with this code:
    sub foo($input) {
        $input.lines.race.map: {
            $/ = G.parse($line);
            $0
        }
    }
since assigning to $/ (which would appear to be a simple fix to get
$/ set so that $0 works), will *still* refer to the $/ of the caller's
scope, aka: $/ of sub foo.

So in the Raku(AST) grammar this should probably be accompanied giving
a block scope its own lexical $/ automatically if it's explicitely
referenced in that scope.

Also cleaned up the code a bit by using fewer nqp ops and more ternaries.