Deprecate IO subs that operate on multiple paths in 6.e
While convienient, being able to pass multiple paths to e.g. &rmdir
is problematic because there isn't a blessed way to handle what to
do when an exception or failure occurs, i.e. should it try to rmdir
every entry even if some things fail, or stop on the first failure?

Previously `rmdir("doesn't exist", "also doesn't exist")` would
return an empty array with no indication of what the failure was.
Yet `("doesn't exist", "also doesn't exist").grep(*.IO.rmdir)
does get failures for these. This difference in behavior is not
intuitive.

This changes subroutines that act on multiple paths by adding a
single path variant, as well as deprecating the multi file variant.
The inclusion of a single path variant is a breaking change, as
these subroutines will now return a failure instead of an array.