Deprecate IO subs that operate on multiple paths in 6.e
While convienient, being able to pass multiple paths to e.g. &rmdir
is problematic because there isn't a blessed way to handle what to
do when an exception or failure occurs, i.e. should it try to rmdir
every entry even if some things fail, or stop on the first failure?

Previously `rmdir("doesn't exist", "also doesn't exist")` would
return an empty array with no indication of what the failure was.
Yet `("doesn't exist", "also doesn't exist").grep(*.IO.rmdir)
does get failures for these. This difference in behavior is not
intuitive.

The idea is that the semantics of providing multiple-path
versions of these subs is a mistake.

This patch changes subroutines that act on multiple paths by adding a
single path variant, as well as deprecating the multi file variant.
The inclusion of a single path variant is a breaking change, as
these subroutines will now return the direct output of the
method form instead of an array that (in some cases) altered
that output.

As currently implemented, this patch leaves the behavior the same
for all previous language versions and adding a deprecation message
in 6.e.

In addition, we take this opportunity to align the syntax of calling
`chmod` to the syntax of `chown`. The aligned candidate is available
as of `6.e`, while the previous signature begins to emit a
deprecation message.