Make sure :chomp is honoured on $*ARGFILES.lines
Or more commonly on the lines() subroutine.  Fixes #5487

The problem is a little insiduous, as $*ARGFILES is an instantiated
IO::Handle object, which basically gets its :chomp setting from when
it is instatiated.

Specifying an undefined value for :chomp will be interpreted as
"take the default from the instantiation".  This should continue to
allow setting the :!chomp state at normal IO::Handle instantiation
and have it passed on to IO::Handle.lines handling.