Implement resolution of non-parametric generics in expressions
Say, `method foo { AGenericClass }` would now return an instantiation of
the class.

Changes required for this to fly:

* `TypeEnv` caching.

  Caching resolves two problems. First, it prevents a class from being
  re-instantiated 2+ times for the same role concretization. Second, it
  provides mechanism for resolving generic types in expression.

* `TypeEnv::LexPad` converted from a mixin role into a subclass of
  `TypeEnv`. This resolves chicken-and-egg problem in role composition.

* `TypeEnv` becomes a value object over its underlying context object

The resolution itself works by installing a special lexical symbol into
the current so called "generic scope" lexpad for each type to be
resolved. The symbol is bound to the type itself. Symbol names are
starting with '!INS_OF_' prefix with full type name attached.

_"generic scope" above is the outermost frame which declares a type
capture. For now this is only role bodies, but in the future routines
with type-capturing parameters are to be added._

Later, when we get a type environment for instantiation, we iterate over
the symbols and bind corresponding type instantiations back to the
symbols in the type environment. When the underlying context is a bare
hash this would only affect further instantiations of the same type. But
when it is lexpad of a closure then the closure gets these instantiations
lexically accessible by all nested frames. The actual resolution happens
by replacing `My::Generic::Type` symbol in the code with lexical
resolution of '!INS_OF_My::Generic::Type'.

Instantiation of referenced generics happens as early as possible,
eagerly. To avoid unnecessary instantiations the compiler installs the
lexical only when the resolution is really needed.

* Added a reference test for class instantiation. The test code is fat
  because we currently don't have any sort of standard API for doing
  these tricks. In the future it would make sense to transfer the test
  to the Roast, as soon as the API is crystallized and published.