Make Match.print more aware of its surroundings
The Match.print method is being use in two ways:
- as the <print> token
- as the method to output $/ or derivative to STDOUT

This was causing problems, e.g. described in #5468.

This commit checks the caller for having an instantiated $¢ in its
lexical scope: this appears to be valid heuristic for determining
whether "print" is being calld as a token, or as a method for writing
to STDOUT.  This condition is then being used to call the correct
method.

Note that this only fixes the issue for "print".  Other cases can
probably be fixed in a similar manner, but are *not* part of this
commit.