RakuAST: remove (almost) all localization support
The only thing remaining is the RakuAST::L10N module which contains
the logic for generating localization modules.  All of the other
files are now being supported in the https://github.com/raku/L10N
repository, allowing localizations to be added / updated *without*
needing a Rakudo release.

The number of installed core modules is now 23 (was 39).