Be more selective when re-parameterizing hashes and arrays
Strip off not all mixing but generic ones only. Normally this would
simply remove `Array::Types`, `Hash::Typed`, or `Hash::Object` mixins
leaving the base classes.

The change is needed because if generic is a parameterization of user
mixin then it's better to preserve the mixin and re-parameterize over
it.

This would strip off any mixins applied to a generic parameterization,
but if this is not desirable then the mixins must take care of the
instantiation by themselves.