Fix arrays and hashes sharing same backing storage
Instantiation of a concrete collection type tries to preserve content of
the generic original by passing the original into constructor call of
the instantiation. While not being a big deal on its own when mapping
1-to-1, this is major problem when the instantiation is used as
attribute container initializer as in this case the object gets
shallow-cloned for each new object of the class-attribute owner.

The solution is to avoid any argumets to the constructor when there are
no elements in the generic instance. In this case the backing storage
would be lazily initialized granting its uniqueness per array or hash.