Prevent possible recursion in archetypes dispatcher
When `is-generic` is a multi and never has been invoked yet sorting of
dispatchees calls `ClassHOW` method `archetypes` recursively. Best if
the call to `is-generic` is ripped away from the dispatcher altogether.

This also simplifies dispatcher's logic. At earlier stages of generic
classes design it did make sense for it to cache the results. Now
caching makes no sense because either the attribute is set by method
`compose`, or some 3rd-party code prevented this to happen
intentionally. This makes it reasonable to treat any class as
non-generic until `$!archetypes` is set; i.e., until `compose` finishes.