Remove defaults from attributes in IO::Handle
In response to https://github.com/rakudo/rakudo/issues/5444 and the
research that vrurg did for https://github.com/MoarVM/MoarVM/pull/1783
this removes the default value setting of the $!chomp, $!nl-in and
$!nl-out to TWEAK, instead being specified on the attribute definition.
By doing this, the default value won't be set internally by the
nqp::attrinited op, which is now suspected of being a possible cause
of random crashes.

If the REA harvester doesn't crash anymore (runs hourly, crashed about
twice a week), then we have a direction to further look into these
crashes.