Handle .extension("foo") on IOs without extension differently (#5512)
Before this commit, "foo".IO.extension("bar") would return "foo".IO
instead of the probably expected "foo.bar".IO.  To get that behaviour
one needs to specify :parts(0) as well.  Which, at least to me, was
a bit of a WAT.

This commit changes that to be more DWIM, so that "foo".IO.extension("bar")
*will* return "foo.bar".IO.  This is achieved by changing the default
of the :parts named argument to 0 if there is no extension yet.

Spectest remains clean.  Ecosystem fallout is expected to be minimal,
as any code expecting to add the extension, would have been expected
to specify the :parts named argument explicitly anyway.