Streamline return value handling some more
Extract the handling of generic types to a new dispatcher
"raku-rv-typecheck-generic", which puts a guard on the type
and then delegates to "raku-rv-typecheck".

Adapt "raku-rv-typecheck" not need to guard on type anymore.

Removes the need for the additional "is_generic" argument to
"raku-rv-typecheck", which saves a few bytes on each callsite.

Document the "raku-rv-typecheck" and "raku-rv-typecheck-generic"
dispatchers.