Remove read lock on MRO
Also replace %!mro by $!mro, as this will be atomically initialized
with an NQPMu, which is easier to check for when inspecting at a
time when the $!mro hasn't been initialized yet.

This appears to save 5 wallclock seconds on spectest and significant
less CPU usage.  This also drops bare startup from .12 to .11 for
yours truly.