Provide native int coercers to Int/Cool
Specifically:
- .int, .int64, .int32, .int16, .int8
- .uint, .uint64, .uint32, .uint16, .uint8, .byte

Note that this is just to provide the API.  The implementation is
definitely hacky: this should probably be handled in the grammar
and in RakuAST it probably will.

This makes the following code work:

    say 256.uint8;  # 0
    say 255.int8;   # -1

And:

    multi a(uint8 $) { dd }
    multi a(Int:D $) { dd }
    a 42.uint8;  # sub a(uint8)

Note however that does **not** make multiple dispatch on different
native sizes work:

    multi a(uint8  $) { dd }
    multi a(uint16 $) { dd }
    a 42.uint8;
    # Ambiguous call to 'a(Int)'; these signatures all match:
    #   (uint8 $)
    #   (uint16 $)

This will need additional work in the dispatcher logic.