Streamline Metamodel::MROBasedMethodDispatch
There seems to be a pattern of building the "all_method_table" for
classes during core compilation.  I wonder if it's invalidating the
cache for each multi method being added (when it shouldn't, as the
only thing in the cache will be the proto).

Also: perhaps .can and .find_method should just use the
all_method_table cache, instead of walking the MRO each time