Use "self" as shortcut for $target.HOW
Although technically it *is* possible to call a method on a
xxxHOW class where the target does not have the invocant as
its HOW, it would really naughty to do so.

So use "self" as a shortcut for $target.HOW where possible.
This should at least reduce the bytecode size.

This also makes naughtiness a DIHWIDT.