Streamline name / ver / auth / api setting
All of the methods that do this, accept these settings as named
arguments.  So add a "set_identity" method in the Versioning role
that will set the appropriate attributes from the info of the
named arguments.

This also centralizes the handling of unnamed classes, making sure
that any counter used is updated threadsafely, so that we can not
get two classes with the same (anonymous) name.

Also change the auth and api to strings, as they will always be
strings.