Change the few cases of if foo -> int bar {
Although this is pretty normal in Raku, it only occurred twice in
the NQP part of Rakudo.  And the few other times I used that construct
I got *very* weird runtime errors.  So even though NQP allows the
syntax, I'm not convinced it actually codegens correctly.

So, it is a bit superstitious, but it felt an easy decision to make.