Streamline Block in bootstrap
Specifically should now be smarter about handling the case where a
block is cloned and it has phasers.

Sadly, the information about which phasers are actually available
in a block, is lost: so we need to check at runtime *every time*
whenever a Block is cloned.  This needs to get smarter.

But for now, this appears to make "test" and "spectest" sigificantyly
faster.