Cache dispatch info on Routine object
The dispatch order logic basically stores a list of hashes with
condensed information about each dispatchee.  And this is kept with
the proto, to allow for faster "find_best_dispatchee" logic.

However, the condensed information about a candidate was re-created
*every time* the dispatch order lists needed sorting again (e.g.
when a new dispatchee was added, or because it didn't survive
precompilation.

This commit adds a $!dispatch_info attribute to the Routine object
which keeps the hash with condensed info.  The Routine.dispatch_info
method will build this information if it hasn't been created yet.

This scrapes off 3 CPU seconds for "make test" for me (1.5%) and
15 CPU seconds off of "make spectest" (1.2%) for me.