Work in progress on integration of ProtoInfo class
The problem is that at the first use of "is rw" on a Routine in
the setting, the dispatch fails.  However, at that early in the
setting, basically nothing works yet, so debugging becomes *very*
hard.  After spending 2 days on this, I decided to leave this
work in a branch for the time being, until more clarity is
obtained.