RakuAST: ignore =finish if there is no code
The =finish construct only makes sense in code.  If it is being used
in a RakuDoc context only, it tried to attach the text to the CompUnit
when there was none, because there was no code.

This now just ignores the =finish text if there is no CompUnit to
attach to.  Potentially, it could also be converted to a RakuDoc,
but this does not actually appear to make sense (yet).