Add multi disambiguation for positional scalars
Until now there has been no way to capture the use of $[]
(as opposed to []) by the caller of a multi routine.

For example:
    multi sub($a where *.elems > 0) {...}
    multi sub(@a) {...}

    a $[1,2]

In this case, the @a candidate always wins (the where clause is never
tested).

Or consider:
    multi sub(Array $a) {...}
    multi sub(@a) {...}

    a $[1,2]; # goes to first
    a [1,2];  # also goes to first

Finally:
    multi sub(Positional $a) {...}
    multi sub(@a) {...}

    a $[1,2]; # failure due to ambiguous dispatch

This patch addresses the issue by removing the ambiguity from this
third example. By specifying a scalar argument to be of type Postional,
it will "win" against an @ argument in the same place in a competing
multi signature.

By taking this approach, it is hoped that the fallout to the ecosystem
will be minimal as it is unlikely that anyone has intentionally shipped
multis that fail to dispatch due to ambiguity.

Note: This patch does not address signatures where the positional scalar
utilizes a where-guard. Since the where-guard already breaks the tie that
is at the heart of this patch, it is much more likely that such a change
would adversely affect the ecosystem.