Add splice Positional $new candidates
We infer the intent of the user in this case to be "insert
this array as a single element".

Thus we pass it to the actual array candidate as a new
array containing a single element that as been "listed"
via the common resolution to the single-arg rule
(ie, `[$single,]`).