Make X::Syntax::Number::LiteralType a subclass of X::TypeCheck::Assignment
X::Syntax::Number::LiteralType gets thrown when we detect at compile
time that an assignment of a literal number will fail due to a type
mismatch. That's exactly the situation X::TypeCheck::Assignment is meant
for. The LiteralType exception just covers a more specific scenario and
will give a hopefully better error message. Thus it seems appropriate to
make it a subclass of X::TypeCheck::Assignment.

Actually in hindsight LiteralType should not have been an X::Syntax
exception at all, since there's nothing wrong with the syntax. But there
are spectests testing for this specific type, so we'll leave it as it is
for now.

This change gives us the flexibility to throw the more targeted
exception type without failing existing spectests.

ab5tract++ for the idea!