RakuAST: fix NYI BUILDPLAN error on typed hash attributes with defaults
A shape on an attribute should only lead to creation of an initializer
thunk for array attributes as there the shape directly influences the
values. On hash attributes we require a value type for generating an
initializer.

Fixes "Defaults on compound attribute types not yet implemented" on
class TypedKeyHash { has %.b{Str} is rw = o => 768; }