Fix "No exception handler located for catch" trying to report method-not-found
$.invocant is not just calling the invocant method but implicitly also
calls .item on the return value. This can fail if $!invocant does not
have that method (e.g. when it's a generic type object) which leads to a
method-not-found error trying to report the method-not-found error only
that there is no more exception handler to deal with it.

Since this code is already the fallback for when we failed to construct
a more elaborate error message, try to make it as safe as possible by
accessing the attributes directly instead of going through accessors.