RakuAST: fix instantiation of generic attributes when applying roles
The logic for detecting generic attributes that need instantiation when
applying a role relies on the properties of the package set for that
attribute. Previously this package was always the package the attribute
was declared in, but for roles it has to be a Generic that will be
instantiated for the class consuming the role, i.e. $?CLASS. After all
the role merely describes what attributes will be added to the consuming
class.

Fixes my role R[::T] { has $.v is default(T); }; class Foo does R[Int] { };
dd Foo.new.v.^name; giving T instead of Int