RakuAST: Fix block topic not being raw enough
In $_ = 1 for @a we want to be able to write to $_. For this we added
the "is raw" trait to the generated signature for the expression thunk.
However this trait was ineffective as it never got applied. The
signature gets generated after BEGIN time effects have already happened.
Fixing this is not easy as to run BEGIN we need a resolver and context
and none of the functions involved get passed both.

Instead of relying on that trait for setting the appropriate flag, we
can just tell the parameter that it's target will want the SIG_ELEM_IS_RAW
to be set.