RakuAST: fix symbols declared in role bodies disappearing for the resolver
After we finish parsing a role we replace the generated Block with a
fully fledged Sub by moving over the block's body. The Block however is
also a lexical scope and as such carries declarations that we also have
to move to the new Sub object.

Fixes role R[::T] { package G { class A is Array[T] {} }; has @.a is G::A; }