Add dispatch disambiguation for 'is item'
This allows for multi routines to distinguish between itemized arguments.

For the sake of simplicity, the parameter trait does not add any new
ParamTypeChecks to the parameter. Instead it is only used as a marker
for disambiguation in multiple dispatch:

    multi sub a(@a is item) { "itemized!" }
    multi sub a(@a) { "array" }

    a [1,2];  # "array"
    a $[1,2]; # "itemized!"

This also works for associatives (${} / {}) and lists ($() / ()).
Sigil-less parameters with Associative or Positional types can also
be disambiguated by 'is item'.

It also works on named parameters. Note that nested-parameters are
currently not parsing with traits at the moment. When they do, this
code will require some adjustment as currently we only use one name
in our name-to-type lookup. Since there was no way to test that it
would work, it has been left for later.