RakuAST: initial stab at adding DEPARSE / raku support
for RakuAST::Role::ResolveInstantiations.  Sadly the effect of having
a RakuAST class without clear source interpretation, makes it a lot
harder to create a sensible source representation that does not have
visual side-effects depending on whether or not there are other
statements in a statement list  :-(

At least .raku and .DEPARSE now do not die with "unimplemented"
errors anymore.