Better fix for additional statements in roles
as done in ea0b0d15753862ad7.  Take the approach of hiding the
additional statements from view, rather than actually provoding
.DEPARSE and .raku support for them.

This is really a temporary measure, as this will be one of the cases
where one can **NOT** use the .AST of a role as a template for building
roles programmatically.  So this *will* need some additional thinking
before this can be unleashed onto the world in a official language
version release.