Fix Inline::Perl5 breakage
By extracting a condition out of a loop, it became possible to call
"language_revision" on a NQP role.  Since language revisions are a
HLL thing, this caused a problem for Inline::Perl5, because it has
its own ClassHOW written in NQP.

Previously, this was impossible because the method would only get
called if a submethod was encountered, which is also a HLL thing.

Fixed by assuming a lower language level if the "language_revision"
method can not be called.