Ensure that candidates carry itemization reqs
Since we already visit all of the parameters here, we
can avoid type and flag checks of these same parameters
in the dispatch code by providing an `item_disambiguation`
flag in the dispatch_info hash that we access later in
the multi dispatch logic.

This is a relatively heavy intervention into the way that
dispatch normally works. It is all to workaround the current
reality of 'callmethod' and 'callstatic' having a bug whereby
the argument passed to ParamTypeCheck is already decont'd.

Since the entire point of `is item` is to dispatch based on
containerization, that's clearly a blocker.

So, in the case of a named parameter with an in `is item` trait
we cannot simply rely on the VM binder to check the candidate's
suitability. Instead we pass the type information of `is item`
named argumants along with the rest of the dispatch_info hash.