Streamline ModuleLoader (Part 4/N)
Move setting up of known symbols out of the loop.  It's unclear
why this was done inside the loop iterating over all source
symbols to be merged.

This is make (spec)test clean, but done in  a separate commit
anyway as to later be able to better bisect should this change
affect the ecosystem.