Use low-level invokability test
Instead of testing for HLL Callable alone.  This fixes the issue of
having low level code blocks in a BUILDPLAN causing havoc, as found
with problems in AttrX::Mooish and possibly Red.

This is probably caused by the removal of some nqp::hllize() calls
in the recent streamlining.  Rather than reverting those changes,
it feels better to do this change, as it also has a significant
effect on performance in some cases (e.g. during "make spectest").