Generate Rakudo Flavor into the actual code
Adapt tools/build/gen-cat.nqp so that it will look for the string
 #RAKUDO_FLAVOR# in the source being processed.  If found, it will
look in the environment for the RAKUDO_FLAVOR variable.  If that is
set, then it will replace the #RAKUDO_FLAVOR# target with a
"' $flavor'".  If it is not found, it will generate an empty string
(as in "''") there.

This also adds the #RAKUDO_FLAVOR# string to Compiler.nqp's version
information, so that --version will show the flavor.

This grants the wish of the Rakudo Star people to be able to mark
their version of Rakudo with an additional "Star" label, by making
sure that the environment variable RAKUDO_FLAVOR is set to "Star"
**at the moment that Rakudo is being compiled**.

This also should fix any objections with users potentially getting
confused by which version of Rakudo they are using, as the flavor
is "burnt into" the actual binary code of the executor, and thus
cannot be changed at runtime.

Additionally, this should allow the flavor to become available in
other areas, perhaps in the form of a $*RAKU attribute, burnt into
in the same manner.