Add worry for WhateverCode on LHS of smartmatch
Smartmatch only curries `Whatever`, so any user that provides
a `WhateverCode` expression as an LHS is likely to encounter
a surprising failure (when compared to the behavior of most
other infix operaators).

In order to be proactive and helpful to such users, we emit
a descriptive worry in this scenario.