RakuAST: re-work currying to run bottom-up
BEGIN handling works bottom-up, i.e. an expressions inner-most parts are
processed first. Previously we had an additional callback for top-down
handling. However we shouldn't actually need that. For whatever currying
(handling of expressions like * + 1) we previously relied on this
top-down run to have a the outer most node in the expression be the
driver of turning that expression into a callable.

This change returns to the original bottom-up approach. I.e. we start at
the inner most node of the expression and when one of its operands is a
Whatever, we curry this. Then the parent checks its operands for
Whatever or curried sub-expressions and it also curries. Curried sub-
expressions get returned to their un-curried state. Thus the curry is
moved up the tree. This keeps all tree manipulations local, i.e. we only
have to touch an operation's operans.

The difficulty in all of this is that we need to keep track of all the
Whatevers we replaced and ensure that the generated Code object's
arguments are bound to the right WhateverCode::Argument nodes we
replaced them with. We get around this difficulty by leaving the part of
generating the curry's signature up to the last step.

During the whole curry/uncurry dance we just replace Whatever nodes we
find with WhateverCode::Argument. Then we do one depth-first-search pass
over the whole subtree to find all those arguments in the order the
corresponding Whatevers appear in the source code. Thus we can just
number them and use them as ParameterTargets in the curry's signature.