RakuAST: replace a block's signature before parsing the blocks body
Placeholders need to check whether they have been explicitly declared as
parameters to know whether they need to add a declaration by themselves.
For this check to actually find the existing parameter, the block's
signature has to be installed already, so do this right after we created
the signature instead of after parsing the block.