Fix mysterious failures when class construction process order changes
For creating a class' full method list (including inherited ones) we
need an up-to-date MRO (i.e. the linearized inheritance list). This MRO
is getting cached. While we already re-calculated this cache when the
parent list changes, a window of opportunity was left open between
method cache getting updated due to a method getting added and the MRO
getting calculated due to a parent getting added.

Previously we were just lucky that all compiler frontend implementations
constructed classes in an opportune sequence that left us with a correct
method cache. However that was not guaranteed. Try to fix this by
invalidating the MRO cache whenever we invalidate the method cache.

This bug was particularly well hidden because we only use the full
method cache when we try to find a method in a megamorphic call site.