Make sure all expected identifiers are available
In the streamlining of NativeCall, quite a few identifiers were
not exported and/or not available in places where some modules
(such as NativeLibs) expect them to be.  This did not lead to
any (spec)test failures, but it did lead to some breakage in the
ecosystem, as reported in:

  https://github.com/rakudo/rakudo/issues/5563
  https://github.com/rakudo/rakudo/issues/5567

This commit restores the exporting and availability of identifiers
from before the streamlining.  And adds a test to check for the
availability of these identifiers in the expected places.