Fix memory leaking in infinity-bound ...
There's no reason to involve the `righti`
iterator at all. Remove that, et voila, no memory leak
for inifinite sequences.

There are only two possible `righti` for infinities:

    > [Inf].iterator.pull-one
    Inf
    > [*].iterator.pull-one
    1

Neither provides anything of any use to the lazy iteration
of the underlying gather that comprises the real reason
for the `...` sequence to exist.

R#1399 (#1399) and RT#131829.