Allow the use of enum values in MAIN that shadow core types
The issue previously is that we would allow users
to create enums with values that overlap with types defined
in CORE, but then when they would try to pass these values
as arguments to MAIN, it would fail and print the usage
string.

Now when doing our "is it an enum value" dance, we additionally
check for whether what we are looking up is from CORE and, if
so, we compare that to what we find in the GLOBAL scope under
the same name. If these don't match, we favor the GLOBAL
version because that's where enums are installed.

Resolves R#5090 (#5090).