Clarify Range.rand failure cases and fix infinite loop
Previously we would be quite unclear about the specific
reasons for failing to produce a random number from a
range:

    > .rand with NaN..NaN
    Can only get a random value from numeric values

    > .rand with 1..0
    Can only get a random value if the range is positive

    > .rand with 1^..^1
    Can only get a random value if the range is positive

    > .rand with 1..1
    Can only get a random value if the range is positive

Now these produce:

    > .rand with NaN..NaN
    Impossible to get a random from a range with a NaN endpoint

    > .rand with 1..0
    Impossible to get a random number from range containing no values.
    Only the sequence (...) operator supports descension between 1 and 0

    > .rand with 1^..^1
    Impossible to generate random numbers for a range where endpoints are equal

    > .rand with 1..1
    Impossible to generate random numbers for a range where endpoints are equal

Furthermore, we now reject any range with endpoints where the difference
is less than 1e-15. This is to prevent the unfortunate case of an infinite
loop:

    > dd .rand == .max with 1^..^1.0000000000000001

For the excluded endpoint cases, the checks for the rand-generating while
loops will always be true because we collapse 1e-15.(Fat)Rat to 0.0. So
we will loop forever.

For the non-excluded endpoint cases, we always produce the $!min+0.0, which
is the same as the rejected scenario when $!min == $!max (regardless of
exclusion).

Ideally we could rely on $*TOLERANCE for something like this, but it has
no effect on Num.(Fat)Rat. Beyond that, `nqp::rand_n` doesn't seem to
produce anything more precise than `1e-15`. So until both of those things
change, this is a simple way to manage the infinite loop/not-random-just-endpoints
scenario.

This addresses R#1591 (#1591).