Test for ambiguous adverb exception
This is only properly testable in RakuAST, as
base dies in NQP's HLL::Grammar without an exception
class to verify against (#1378).

We could test against the text of the die message, but
my experience has led me to consider that bad practice
and the situation in base is only temporary anyway.