Provide deeper context into `run` failures
After 92bd7e4, we stopped returning an exit code of `-1` for
cases when attempting to `run` a nonexistent command.

This patch returns the previous behavior while also surfacing
the underlying error from `libuv`, which makes it much more
clear to the user what the underlying cause of the `run` failure
is:

    > r 'run "e-not-a-command"'
    The spawned command 'e-not-a-command' exited
        unsuccessfully (exit code: -1, signal: 254)
    (OS error = Failed to spawn process e-not-a-command:
        no such file or directory (error code -2))

R#1590 (#1590) and relies on MoarVM@13598bbfc9.