Provide some fixups to experimental :cached
Addresses R#4666 (#4666), the discussion of which
raised several issues with the previous `is cached`
implementation:

- `.gist` is not appropriate as details may be truncated,
  leading to potential invalid cache matches

- `is cached` is not thread-safe

This patch moves to address both issues by using `.raku` instead
and wrapping it in a `lock`/`unlock` cycle. (`Lock.protect` is
not an option as it hijacks the `callsame`).

Additionally, `Setty.raku` and `Baggy.raku` have been adjusted
to produce sorted contents.