Additional guard when selecting the type for MAIN args
This is required to fix the situation where ::('a') and
CORE::('a') are the same, but GLOBAL::('a') doesn't
match CORE::('a') by virtue of being `Nil` -- the intent
of the guard is to use GLOBAL::('a') when it is "in front"
of CORE::('a') by virtue of being part of a user-defined
enum that shadows something of the same name in CORE.

Since that won't be the case when GLOBAL::('a') is Nil, we
should include this in our guard.

This fixes the situation that broke the `SuperMAIN` module,
which relied on `"True"` becoming `Bool::True`.

Note that it is still possible for users to do something like:

    ## test.raku
    enum Questionable <Nil Bool>;

    sub MAIN(Questionable $thing) {
        dd :$thing;
    }

    > ./rakudo-m test.raku Nil
    # :thing(Questionable::Nil)

R#5548 (#5548).